Assignments/HW4/myproject/pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = ""
authors = ["Temirbolat Maratuly <snake2009003@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
six = "^1.16.0"
filetype = "^1.2.0"
django-debug-toolbar = "^4.4.6"
python-decouple = "^3.8"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
